make sure diag file created even if no cat required

+++ run_gsi_4densvar.sh
@@ -945,28 +945,34 @@ for loop in $loops; do
    for type in $alldiag; do
        count=`ls pe*${type}_${loop}* | wc -l`
        if [[ $count -gt 0 ]]; then
-          export PGM="${execdir}/nc_diag_cat.x -o ${savdir}/diag_${type}_${string}.${adate}_${charnanal2}.nc4  pe*${type}_${loop}*nc4"
-          ls -l pe*${type}_${loop}*nc4
-          nodecount=$((nodecount+1))
-          if [ "$machine" == 'theia' ]; then
-             node=`head -$nodecount $NODEFILE | tail -1`
-             export HOSTFILE=hostfile_${nodecount}
-             /bin/rm -f $HOSTFILE
-             n=1
-             while [ $n -le $nprocs ]; do
-                echo $node >> $HOSTFILE
-                n=$((n+1))
-             done
-             echo "contents of hostfile_${nodecount}..."
-             cat $HOSTFILE
+          if [[ $count -eq 1 ]]; then
+            # just one file, no cat needed (just copy it)
+            file=`ls -1 pe*${type}_${loop}*`
+            /bin/cp -f $file ${savdir}/diag_${type}_${string}.${adate}_${charnanal2}.nc4
+          else
+            export PGM="${execdir}/nc_diag_cat.x -o ${savdir}/diag_${type}_${string}.${adate}_${charnanal2}.nc4  pe*${type}_${loop}*nc4"
+            ls -l pe*${type}_${loop}*nc4
+            nodecount=$((nodecount+1))
+            if [ "$machine" == 'theia' ]; then
+               node=`head -$nodecount $NODEFILE | tail -1`
+               export HOSTFILE=hostfile_${nodecount}
+               /bin/rm -f $HOSTFILE
+               n=1
+               while [ $n -le $nprocs ]; do
+                  echo $node >> $HOSTFILE
+                  n=$((n+1))
+               done
+               echo "contents of hostfile_${nodecount}..."
+               cat $HOSTFILE
+            fi
+            sh ${enkfscripts}/runmpi 1> ${current_logdir}/nc_diag_cat_${type}_${string}_${charnanal2}.out 2> ${current_logdir}/nc_diag_cat_${type}_${string}_${charnanal2}.err &
+            #sh ${enkfscripts}/runmpi 1> nc_diag_cat_${type}_${string}.out &
+            if [ $nodecount -eq $totnodes ]; then
+               echo "waiting... nodecount=$nodecount"
+               wait
+               nodecount=0
+            fi       
           fi
-          sh ${enkfscripts}/runmpi 1> ${current_logdir}/nc_diag_cat_${type}_${string}_${charnanal2}.out 2> ${current_logdir}/nc_diag_cat_${type}_${string}_${charnanal2}.err &
-          #sh ${enkfscripts}/runmpi 1> nc_diag_cat_${type}_${string}.out &
-          if [ $nodecount -eq $totnodes ]; then
-             echo "waiting... nodecount=$nodecount"
-             wait
-             nodecount=0
-          fi       
        fi
    done
